requirements.txt
boto3==1.13.4
botocore==1.16.4
Django==2.2.10
django-crispy-forms==1.9.0
docutils==0.15.2
jmespath==0.9.5
pip==20.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
setuptools==40.6.2
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
